test: split bug-report tests into their co-located test files
Move the 15 SFT/DPO/ORPO bug-report tests out of the catch-all
test_sft_dpo_bug_report.py and into the existing test files that own
each module, so each test lives next to the code it exercises:

- test_batched_losses.py: DPO/SFT loss off-by-one (response_start slice).
- test_dpo_loop.py: epoch 0 + cross-epoch shuffle, total_steps rounding,
  and load_preference_dataset validation (fail-fast on malformed rows).
- test_orpo_loop.py: structural check that the per-epoch shuffle is seeded.
- test_sft_loop.py: data-pipeline padding and eval carve-out shuffle.

No behavior change; relocating tests only. Full cookbook unit suite
(415 tests) still passes with 0 regressions.

Made-with: Cursor
